Don't attempt to handle command lines involving @foo
The @ switch can be used to pass additional options and files to the
compiler via a file instead of using the command line. This maybe used in case
the maximum command line length was reached.

We don't know how to look into (or interpret) the referenced file yet, so let's
just discard such command lines.