Correctly forward return code of cl.exe
We used to call sys.exit() with the return value of the invokeRealCompiler
function. However, this function returns a tuple and not a plain int. Passing
a tuple to sys.exit() makes it just exit with '1'.

Fix this by passing the first member of the tuple (the actual return
code). Patch based on a contribution by Chris Hillery.