Use FileLock class to fix race conditions when accessing the cache
github issue #6 reveals some race conditions in the code which can happen if
the clcache script is run multiple times in parallel. One instance may be busy
with cleaning the cache while another instance adds data, or the like.

Fix this by using a lock file. This is probably not the most efficient
solution, but I'd like to be correct in the first place.