Trying to lock cache directory consistently
Executing clcache.py in parallel might cause race conditions when
accessing the cache directory. All access to the cache directory
needs to be synchronized, so we need to implement some kind of
locking. Maybe I should rather use a simple SQL database or the
like which gets me ACID semantics for free...