Made findCompilerBinary() more robust in case we were frozen using py2exe
When using py2exe, it may well be that the user renamed the resulting
binary to 'cl.exe' so that it's a drop-in replacement. In that case,
we need to be more careful when trying to find the real compiler binary
because it's very well possible that the first cl.exe binary we find
is ourselves. If we used that, we'd invoke ourselves recursively.