Use Windows Mutex to implement object cache locking
This works better in case clcache gets terminated abnormally, in which case you
might end up with stale lockfiles. A mutex won't have this problem, the handle
is released when the process terminates.

A mutex might also be faster since there's no disk access anymore.

This patch is mostly courtesy of Vyacheslav Chigrin <vchigrin@yandex-team.ru>,
he did most of the work as part of commit
72094faa06eaba468c2cafc57d9137d96ff3aadd.