When cleaning the cache, clear the oldest object files first
Vyacheslav Chigrin <vchigrin@yandex-team.ru> pointed out that as code evolves,
old object files are less likely to be reused. I tend to agree, so let's sort
the list of object files so that those with the oldest access time are first in
the list, so that we delete them first.