Fixed grave bug when computing cache key
It turned out that when computing the cache key (and when invoking
the preprocessor!) we always droppd the first item from the given
'commandLine' argument. This got broken by commit
ca1c51404e3835b2e9f754161071cce238c073e3 which caused 'commandLine'
to only contain the arguments to the compiler, not the actual
compiler itself.

The cache key computation code still assumed that the first element
of the command line is the compiler binary, so it skipped that.

This bug not only caused some builds to break (because e.g. important
include paths were missing!), it also means that the first compiler
argument was not taken into consideration when computing the cache
key. So compiling the same soruce twice, but with slightly different
compiler invocations which only differ in their first argument, would
incorrectly result in a cache hit.