Fix compile error when clcache is used with precompiled headers
This passes all calls including /Yc and /Yu to the real compiler,
allowing the use of clcache in a mixed environment, where compilation
units without precompiled headers use the cache and compilation units
with precompiled headers don't.

A test project for precompiled headers was added to tests/precompiled-
headers. With this test project it was possible to reproduce the compile
failure.

Disclaimer: The author of this sample project has no clue about
precompiled headers, does not use them in production code and has
no interest in doing so in the future.