Fixed locking strategy to avoid corrupting cache stats
Concurrent invocations of clcache.py would easily break the cache
statistics. The reason for this was that in some cases, a
read/modify/write session on the statistics was not guarded by a single
lock, so two clcache.py processes A and B which try to bump some value
(e.g. number of cache entries)might have executed a sequence like

  A read value 5
  B read value 5
  A write value 6
  B write value 6

Fix this by rewriting the locking strategy. Locking is no longer done at
the lowest level (since such locks do not compose) but at the toplevel
in all relevant areas:

- When printing the cache statistics
- When resetting the cache statistics
- When changing the cache configuration

Furthermore, when processing a compiler request, the lock is acquired at
least once in order to:

- Check if the invocation can be cached at all
- Check if there is a cache entry

Then (without the lock being held!) the compiler is invoked. Finally,
depending on the invocation and the configuration settings, a
post-processing function is invoked to update the bookkeeping. This
function also takes care of holding the lock as needed.