Added initial tests for clearing the cache (and make them pass)
This adds just some basic sanity checking for clearing the cache:
clearing should be idempotent, i.e. repeatedly clearing the cache
should work and maintain the same postcondition. Also, verify that
clearing the cache actually removes the objects (and thus reduces the
cache size) but otherwise leaves the cache statistics untouched.

Even these simple tests showed an oversight in the code for clearing the
cache: an undefined identifier 'fn' was used, but 'o' was meant. So
let's fix that.