Make test failures break the CI build again
The recent patch for executing the tests via py.test introduced a
PowerShell wrapper script which didn't forward the return code back to
AppVeyor as intended. This caused non-zero exit codes by py.test (which
indicate test failures) to not be visible to AppVeyor, hence the CI
build succeeded. We do want the build to fail if the test fails though.

This patch does two things:

  1. It splits the PowerShell scripts into two pieces such that both the
  unit as well as the integraiton tests are run separately and then
  2. remembers the exit code of the py.test invocation and makes the
  PowerShell script return that.

Splitting the script is done such that the test results are always
uploaded, even if some test fails.

Doing even an 'exit 0' in a PowerShell script is considered a build failure
by AppVeyor, it seems. The page

  http://help.appveyor.com/discussions/problems/1322-powershell-exit-code-is-failing-build

suggests that this can be worked-around by just exit'ing if the return
code is non-zero.

Fixes #106