Print real command line as array, not as string
Aguments are passed to Popen as a list and it is important to understand
where each argument is starts and ends.

Before this ['cl.exe', '/c', '/nologo'] and ['cl.exe', '/c /nologo']
have been printed the same way, but the later one is invalid. This
change lets us detect this kind of bugs.