Rewrote splitCommandsFile() in terms of a new CommandLineTokenizer class
Over the last couple of years, there have been numerous bug reports
about how clcache fails to parse response files correctly. This was
always handled in a somewhat ad-hoc manner: we tried to patch the code
such that the newly discovered case works, without breaking the old
ones.

Being dissatisfied with this situation, I looked at how other projects
parse response files and became aware of some code in the LLVM project
solving this exact issue. In particular, the "cl::ExpandResponseFiles"
function[1] works a lot like our code, except that it uses a separate
and more complex tokenizer function "cl::TokenizeWindowsCommandLine"[2]
to actually parse the response file.

The latter appears to implement the rules documented in the MSDN article
"Parsing C++ Command-Line Arguments", which outlines a couple of rules
which significantly diverge from the rules for parsing shell command
lines.

The new 'CommandLineTokenizer' class introduced by this patch is a port
of the LLVM function 'cl::TokenizeWindowsCommandLine', slightly
exploiting the fact that functions in Python are nicer first class
citizens (so a 'state' variable is not needed).

A couple of existing tests fails with this, but I believe the tests may
just have wrong expectations based on the lack of understanding for how
response files are parsed.

[1]: http://llvm.org/docs/doxygen/html/namespacellvm_1_1cl.html#ab625bb6f7f0e37e795cb7833de9ac079
[2]: http://llvm.org/docs/doxygen/html/namespacellvm_1_1cl.html#a3b42fd69f84c0ceef44857e925613ee4
[3]: https://msdn.microsoft.com/en-us/library/windows/desktop/17w5ykft(v=vs.85).aspx