Fix float vs int mismatch with Python 3
As explained on http://stackoverflow.com/a/183870/91757:

  In Python 3.0, 5 / 2 will return 2.5 and 5 // 2 will return 2. The
  former is floating point division, and the latter is floor division,
  sometimes also called integer division.

  In Python 2.2 or later in the 2.x line, there is no difference for
  integers unless you perform a from __future__ import division, which
  causes Python 2.x to adopt the behavior of 3.0

I only actually used this code with Python 2.7, so never noticed the
issue.

Just using '//' with either Python version would have done the job, but
I decided to also enable __future__.division here to make sure that
similiar issues get caught earlier in the future, even with a Python 2.x
installation. As it is however, there is no other place where we are
dividing values, so there should be no immediate effect of importing
__future__.division.