Fixed TestSplitCommandsFile.testQuotesAroundArgument and .testDoubleQuoted
The Parsing C++ Command-Line Arguments page[1] explains

  A string surrounded by double quotation marks ("string") is
  interpreted as a single argument, regardless of white space contained
  within. A quoted string can be embedded in an argument.

and gives the example that the string

  a\\b d"e f"g h

is to be parsed into three arguments

  a\\b
  de fg
  h

I.e. 'd"e f"g' is one argument, but the quotes are stripped. Hence,
'/Fo"foo bar"' should be one argument '/Fofoo bar'.

Furthermore, another example is that

  "abc"
  d
  e

i.e. quotes enclosing the entire argument are redundant (in case the
argument has no whitespace characters) and are dropped.

[1]: https://msdn.microsoft.com/en-us/library/17w5ykft.aspx