Fixed TestSplitCommandsFile.testBackslashBeforeQuote and .testVyachselavCase
The Parsing C++ Command-Line Arguments page[1] explains

  A string surrounded by double quotation marks ("string") is
  interpreted as a single argument, regardless of white space contained
  within. A quoted string can be embedded in an argument.

  [..]

  A double quotation mark preceded by a backslash (\") is interpreted as a
  literal double quotation mark character (").

...with the latter rule being precised in

  If an even number of backslashes is followed by a double quotation
  mark, one backslash is placed in the argv array for every pair of
  backslashes, and the double quotation mark is interpreted as a string
  delimiter.

I.e. the two-character sequence \" is not *always* a literal " but only
so if it's the suffix of a (potentially empty) sequence of an even
number of backslashes.

Hence, the first " in

  /Fo"C:\out dir\"

is starts an argument (but is dropped from the output) but the second one is a
literal " which *does not* end the argument, hence

  /Fo"C:\out dir\" /nologo

is actually just one argument

  /FoC:\out dir" /nologo

[1]: https://msdn.microsoft.com/en-us/library/17w5ykft.aspx