Added same 'trailing backslash' tests
The old ones were somewhat pathological because they clearly didn't do
what the caller intended. The \" sequence caused a literal " to be
parsed, so the string argument never got closed (only at the end of
input).

Let's add some sane tests which test what was mentioned. Alas, the
notation for ending a raw string in Python with a backslash is somewhat
awkward, as explained by

  http://docs.python.org/faq/design.html#why-can-t-raw-strings-r-strings-end-with-a-backslash