Make local variable `sourceFiles` plural; let analyze() return list
analyze() returned a list of source files or None in all cases but
AnalysisResult.Ok. This commit ensures that the second return value is
always a list. This is important because the second return value is
stored in a local variable at a point where we don't yet know the
analysis result. This local variable now has a proper plural name
`sourceFiles`.