Test parsing of arguments starting with backslash
Looking at the coverage of the TestSplitCommandsFile test case showed
that the only statements in the CommandLineTokenizer which were not
covered yet were the ones handling the case that some argument *starts*
with a backslash.

A real-life scenario of this situation is when source files are in the
root directory (not uncommon when using the 'subst' command), e.g.

  C:\>subst X: C:\Users\Frerich\src\fooproject

  C:\>X:

  X:\>cl \foo.cpp

Let's add a test for this, so all statements have been covered according
to Coverage.py.