Be more explicit about creating a new compiler command line
Quantifiedcode suspected that we merely want to extend the 'cmdLine'
list, and thus suggested using extend() or +=. The code indeed doesn't
make it clear that we really want to create a new list. so let's try to
be more explicit by setting cmdLine to a new list and then extending
that. I hope that Quantifiedcode understands this intention now. If it
still complains, we may need to introduce a new variable name.