Drop base class in favor of class setting
Quantifiedcode (maybe incorrectly so) complained that there were methods
in a TestCase which were neither setUp() nor tearDown(), nor did they
start with "test_". I'm not sure whether there should be an exception to
this rule for constructors, but in any case it seems that the base
classes is not needed.

The documentation of unittest.TestCase.longMessage explains that

  The class setting can be overridden in individual tests by assigning
  an instance attribute to True or False before calling the assert
  methods.

...i.e. the setting is global, but can be overriden per test case. Since
we always set it to True via the BaseTest base class, let's just drop
the class altogether and set the class setting