Rename sourceFiles -> inputFiles in analyze()
When `inputFiles` is assigned, we do not know if the list of files is
source files or for example object files (from `/link file.obj
file2.obj`). Only at the end of the function when no AnalysisyError is
raised, we know for sure that we are dealing with source files. Thus we
can assign the functions result to `sourceFiles` in
`sourceFiles, outputFile = CommandLineAnalyzer.analyze(cmdLine)`