Trying to make flagged coverage reports more useful
It seems to me that right now, the flagged results actually don't work:
both labels seem to correspond to the same coverage information.

Maybe my understanding of how codecov works wasn't quite correct: I
assumed that multiple (potentially orthogonal) reports are to be
uploaded with different flags. That doesn't seem to be the case - maybe
we need to generate the coverage data incrementally? The --cov-append
switch to py.test implements this, let's see how it goes.