Work around quirk in current pylint releases
pylint somehow doesn't notice that 'buffer' is a member of both
sys.stdout as well as sys.stderr, so it complains - which makes the CI
break. This was reported at

  https://github.com/PyCQA/pylint/issues/1004

In the meanwhile, let's reintroduce the 'printBinary' function and use
it as a sole place to disable this particular pylint warning.