Factoring our license terms to separate LICENSE file
...instead, just briefly mention that this is BSD 3-Clause licensed and
that the exact text of that license can be found in the LICENSE file.
This shortens the other files a bit. I also added the license to the
setup.py while I was at it.

As a consequence, both Tim and myself are now copyright holders on the
test code I think since so far only Simon had the copyright declared but
with a central LICENSE file we have a central list of contributors. I
hope Simon is okay with this. :-)