Optimized cache misses in direct mode if CLCACHE_BASEDIR is set
In case CLCACHE_BASEDIR was set, the parseIncludesList() function would
automatically make all include paths relocatable by replacing the base
directory with a magic placeholder. In postprocessNoManifestMiss()
however, we would immediately reverse this operation by expanding the
placeholder again.

The only reason why we may want a placeholder is to store relocatable
paths in the Manifest object, so let's do that there and otherwise deal
with full paths.