Sort includes set
Includes are uniquely stored in a set, which avoids redundant file reads
when restoring in direct mode. This is good.

Now this set was converted to a list with no defined order with the
potential that {'a', 'b'} is converted to ['a', 'b'] one time and ['b',
'a'] another time. This leads to different manifest hashes for the same
includes set, which is undesirable.