Introduced ManifestSection class
Objects of this class represent a 'manifest section', i.e. one of the
256 possible subdirectories of the 'manifests' directory. A section
provides the API for getting the path to a manifest file as well as
setting or getting a manifest.

So far, a section object is only used internally in ManifestsManager,
and all the API forwards. I plan to inline the forwarding functions in a
subsequent step such that callers are ported to working directly with
sections.

The long term goal is to be able to perform section-specific locking
such that we don't need to use the global lock for working with
manifests.