Centralized opening manifest section in processDirect()
This moves the access to the manifest manager in order to get a handle
on the right manifest section out of postprocessHeaderChangedMiss() and
postprocessNoManifestMiss(), back into processDirect().

This means that processDirect() is now the only place where a specific
manifest section is opened, and then that section object is passed
around. This should make it easy to introduce a lock specifically for
the used manifest section.