Moved section-specifc cache accesses from ObjectCache to CacheSection
This patch moves the definition of various methods from the ObjectCache
class to the CacheSection class. The plan is that section-specific
accesses (which only are relevant to a single section of the cache) only
require holding a single CacheSection object such that in a later step,
we can lock just that one section to improve concurrency. This is a
means to address the poor concurrency during cache hits reported in GH
issues #160.

The API in the ObjectCache class is still available so that the callers
don't need to be adjusted just yet. All the methods are merely
forwarding to the respective cache section though.