Introduced new 'CompilerArtifacts' named tuple
This is meant to bundle all the compiler artifacts we hold in our cache.
Right now, this is the path to the object file generated by the compiler
(we yield a path instead of the raw binary data to allow efficiently
copying the data) as well as the stdout and stderr output of the
compiler.