Removed cachedCompiler(Output|Stderr) from CompilerArtifactsSection
These methods were only used in CompilerArtifactsSection.getEntry().
Instead of making them private, let's discard them right away in favor
of a shared _getCachedCompilerConsoleOutput() method which takes a file
name to read (and decode) console output from. A consequence of this is
that trying to read the stdout output of a cache entry will no longer
raise an exception in case the file does not exist but rather yield an
empty string, much like for stderr. That should be okay.