Fixed /MP causing a lot of cache misses
I noticed that when compiling two source files individually and then
compiling both of them in one go while specifying /MP would not cause
any cache hits. It turns out that this is because the /MP switch
contributes to the hashed command line, and when reinvoking itself
recursively, clcache does not strip the /MP switch for the individual
child processes.

Let's fix this by stripping the flag in both direct as well as indirect
mode. I think stripping this flag even in direct mode should be safe
since it doesn't influence the preprocessor. I think that technically,
we could strip a couple more flags in direct mode, though we need to
keep all those relevant to the preprocessor.