Redefined Cache.lock field in terms of section & statistics locks
This reimplements the global 'Cache.lock' lock such that it's defined in
terms of the individual section and the statistics locks. This means
that acquiring and releasing Cache.lock acquires and releases locks for
all sections and for the statistics. This is slower than before
(because it requires acquiring and releasing up to 513 locks) but it
should be only rarely needed - mostly, when cleaning the cache.