Use section-specific locks in favor of global Cache.lock
This (big) patch attempts to improve concurrency by avoiding the global
cache lock. Instead, we will lock only those manifest and compiler
artifact sections which we deal with. The only cases where we need to
synchronize all concurrenct processes is when updating the statistics
file, because there's only a single statistics file. At least for cache
hits this could be avoided by tracking the number of cache hits per
section, too.

To avoid deadlocks, the locks have to acquired in the same order for all
execution paths (the order is defined in Cache.lock, i.e. first
manifests, then artifacts, then statistics). Hence, locking of the artifact
section had to be pulled out of the addObjectToCache() function since
that function was called with the stats locked already - a violation of
the locking order.

Furthermore, we can no longer perform cache.clean() in
addObjectToCache() because cache.clean() acquires the global lock, so
e.g. this sequence of steps was possible in non-direct mode:

1. 'A' locks cache section 'x'
2. 'A' locks global statistics lock
3. 'B' locks cache section 'y'
4. 'A' needs to lock all sections to do a cleanup

At this point, 'B' cannot proceed because 'A' still holds the statistics
lock, but 'A' cannot proceed because 'B' still holds the lock on section
'y'.

This issue is caused by -- from B's vie -- the statistics lock being
locked before a section lock. This must never happen.

At the point addObjectToCache() is called, we already have the
statistics locked and we know that it may be that the cache size limit
was just exceeded, so it's a good moment to determine that a cleanup is
needed. It's not a good moment to *perform* the cleanup though.

Instead, let the function return a flag which is propagated all the way
back to processCompileRequest(). The flag indicates whether cleanup is
needed, and if so, processCompileRequest() will acquire Cache.lock
(which acquires all sections and statistics in the correct order) to do
the cleanup.