Fixed running TestPrecompiledHeaders.testSampleproject
The elements of the CLCACHE_CMD list may contain spaces in case the
source code is stored in a directory with spaces in it. In such cases, a
naive ' '.join(..) does not do the right thing and thus this test
failed.

Fix this by using the somewhat unknown (only briefly mentioned in the
documentation) subprocess.list2cmdline function which takes care of
escaping for Windows. It appears this does the right thing.

This resolves #206.