Adding 'clcachesrv' server script
This little server opens a named pipe and waits for incoming messages.
Each message is expected to be a list of newline-separated paths, and
the message is terminated with a null byte For each path, the server
will compute *and cache* the hash sum and return it.

The server also uses filesystem monitoring API to detect when any of the
files for which the hash sums are cached changes; in those cases, the
hashes will be invalidated automatically such that they are recomputed
on the next try.

In case anything goes wrong while handling a request, we serialise the
exception value and send it over so that the client can throw it; this
usually happens when trying to hash non-existant files.

To implement the asynchronous IO, I went for the pyuv module, a Python
wrapper around libuv. It seems to work well, but unfortunately is not a
trivial dependency. Hence, I moved the server into a separate directory
which is expected to host a virtual environment.