Moved compiler invocation into lock-free zone in processDirect()
This improves concurrency of cache misses by making sure that the real
compiler is only invoked without holding any locks, so the risk of
hitting timeout errors because compilation of an individual file takes
very long is greatly reduced. However, the patch also introduces a race
condition in which two concurrent clcache invocations try to add the
same cache entry at the same time - I'll fix this in a subsequent
commit.