Fixed racy cache writes in processDirect()
Since the real compiler is no longer invoked while holding all relevant
locks, it may be that two concurrent clcache invocations rebuild and
then add the same cache data. To avoid that this causes conflicts when
writing out the data, ensure that the cache entry does not exist already
before trying to add it.

It *appears* that this is only necessary for the compiler artifact
repository -- calling setManifest() on a ManifestSection is apparently
already safe; at least if I'm reading the code correctly...