Moved compiler invocation into lock-free zone in processNoDirect()
This improves concurrency of cache misses by making sure that the real compiler
is only invoked without holding any locks, so the risk of hitting timeout
errors because compilation of an individual file takes very long is greatly
reduced. To avoid that we introduce a race condition caused by two concurrent
clcache processes rebuilding and then adding the same data to the cache, we
make sure to test the artifact section once again after the build to see
whether the cache really needs updating.