Don't assign lambda expressions to variables
QuantifiedCode complains that you shouldn't assign a lambda expression
to a variable but just use a plain 'def', and I tend to agree. However,
let's use a third option here: just as in other test functions, let's
define a "runCompiler" function -- but parametrize it on the include
path. We then pass two lambda expressions to expectHit() which invoke
that function with different paths.