Enable clcache to use clcachesrv
This makes clcache honour a new CLCACHE_SERVER environment variable; if
it is set, clcache will not compute hash sums for files itself but
rather expect that there is a clcachesrv instance running. This should
improve the runtime for cache hits.

A response starting with '!' indicates that an exception occured, and
the rest of the response can be deserialised using pickle to get an
exception value which can be thrown.

What's a bit ugly is that I had to duplicate the pipe name.