Made TestStatistics and TestConfiguration idempotent
It wasn't possible to successfully run TestStatistics.testHitCount()
twice in a row; the test modifies a fixed file and thus violates a
precondition for the next run.

I briefly considered makig the test not expect 0 and 1 but rather n and
n+1, but it's still ugly to create a JSON file in the checkout directory
since Git shows it as an untracked file.

This is also an ugly wart about TestStatistics.testOpenClose() as well
as the two tests in TestConfiguration

Let's just use a temporary file name instead. Alas, the Settings() and
Configuration() constructors cannot be called with a file object, so we
cannot just use tempfile.NamedTemporaryFile(). Instead, we need a file
name, so a helper function gives us that (which introduces a miniscule
chance for race conditions of course, oh well).