Fixed off-by-one in runJobs()
In the course of discussing GitHub issue #239, @TiloW noticed that the
number of clcache instances launched by the runJobs() function was
always one larger than the number of intended concurrent instances. A
simple off-by-one. This patch fixes it.

Alas, I couldn't think of a nice way to create a test for this since
controlling concurrency is quite difficult and I'd rather have no test
than a flaky test.