Do not recommend using pyinstaller with --onefile
It was noted (in GitHub issue #232) that using the --onefile argument
with pyinstaller significantly slows down launching the generated
binary. Indeed, the PyInstaller documentation explains:

  The one executable file contains an embedded archive of all the Python
  modules used by your script, as well as compressed copies of any
  non-Python support files (e.g. .so files). The bootloader uncompresses
  the support files and writes copies into the the temporary folder. This
  can take a little time. That is why a one-file app is a little slower to
  start than a one-folder app.

Startup time of clcache is relevant for common use cases, so let's make
sure that this overhead is avoided by not recommending to use this
switch.