Add atomicRename() and write json files to temporary files before rename
According to documentation and some internet posts os.replace(...) is atomic
 (https://docs.python.org/3/library/os.html#os.replace). This commit changes
the code writing the manifest and the statistics files to write to a temporary
file and then use a new atomicRename() function to atomically replace the old
file with then new contents.

It should avoid corrupting the cache when the clcache process crashes or is
killed during a write, partly addressing the root cause issue #263. Note that
the object files are still not updated atomically.