Re-add touchEntry
Since the manifest is not locked for reading before calling touchEntry it
may have changed on disk. It is needed to lock, read the manifest, touch
and write it atomically. touchEntry(...) also cannot rely on the index to
be valid after read, so it was re-implemented to use the objectKey to find
the correct entry.