Use atomicRename() from copyOrLink
Use atomicRename() to save and restore the object files from the
cache.  It should reduce problems with corrupted object files in the
cache as reported in issue #263.
While it fixes the most important part (corrupted object files) it
still it does not completely fix the problem as the cache entry is
composed of:

 * object file
 * stdout
 * stderr

which should be replaced atomically.