Made clcache source code pass mypy checks
There were four issues which made 'mypy' complain about the clcache
source code:

1. There was a syntax error in a type comment; mypy complained:

  clcache.py:1254: error: syntax error in type comment

It appears that for function types, the arguments need to be enclosed in
parentheses.

2. mypy failed to process type comments because of various meta types
missing; import the relevant types from the 'typing' module, but only if
it's available (it was introduced with Python 3.5)

3. Decorated properties such as

    @property
    @contextlib.contextmanager
    def f():

are not supported by mypy and generate an error. Some Google-query
reveals that this can be avoided by putting '# type: ignore' on the
'@property' line.

4. mypy seems to generate a .mypy_cache directory - let's have Git
ignore that.

With all this in place running

  mypy --ignore-missing-imports clcache.py

passes without any errors.