Fix upgrading pip/setuptools/wheel during AppVeyor CI builds
The CI builds started failing a while ago, printing:

  pip install -U pip setuptools wheel
  ERROR: To modify pip, please run the following command:
  c:\python34\python.exe -m pip install -U pip setuptools wheel

I.e. it sounds as if one should not invoke 'pip' directly for some
reason but rather load it as a Python module. Looking at the YAML file,
it seems we already did that -- right after the problematic 'pip install
-U' call! So let's drop the direct 'pip' call in favor of the existing
'python -m pip ...' call to fix this.