Resolved various items pointed out by pylint
Pylint pointed out that deriving from 'object' is not needed in Python 3
and can safely be omitted. Furthermore, it warned that we're comparing
callables -- which indeed is often not what you want, but in this case
it is. So let's disable that warning.

Last but not least, an incorrect import order was detected in one of the
test scripts.