Clean up adjustments to PATH
I noticed warning messages like

Installing collected packages: mccabe, isort, wrapt, lazy-object-proxy, astroid, pylint
  The script isort.exe is installed in 'c:\python34\Scripts' which is not on PATH.

in the logs of our CI builds. It appears that this warning is caused by
an unintended extra backslash when extending the PATH environment
variable; we ended up prepending

  C:\Python34\\Scripts

to the PATH, i.e. with a double backslash. Let's fix this to get rid of
the warnings.