Fixed invoking codecov.exe in PowerShell scripts
For some strange reason, codecov.exe wouldn't be executed even though
it's directory was on the PATH. Instead, an error message saying

    codecov.exe : The filename, directory name, or volume label syntax is incorrect.
    At line:9 char:1
    + & codecov --no-color -X gcov -F unittests -e PYTHON_INSTALL
    + ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
        + CategoryInfo          : NotSpecified: (The filename, d...x is incorrect.:String) [], RemoteException
        + FullyQualifiedErrorId : NativeCommandError

was printed.

Some tinkering revealed that this appears to be an issue in recent
codecov versions. Codecov v2.0.10 appears to be the most recent version
which is not affected by this issue, so let's enforce that particular
version for the time being.