Drop support for Python 3.4
The CI builds currently fail since installing pypiwin32 fails with:

Collecting pywin32>=223 (from pypiwin32)
  Could not find a version that satisfies the requirement pywin32>=223 (from pypiwin32) (from versions: )
No matching distribution found for pywin32>=223 (from pypiwin32)
Command exited with code 1

It appears that PyInstaller is simply not supported with Python 3.4.
Technically, it might be nice to continue doing basic sanity checks for
Python 3.4, but Python 3.7 was released a while ago so I'd rather not
bother tinkering with the appveyor.yml while to disable building the
installer with Python 3.4 but rather drop 3.4 support completely and add
support for 3.7 instead.

Supporting Python 3.7 (thus sticking to the scheme of supporting the
three most recent Python releases) would be nice, but alas, it appears
that PyInstaller 3.3.1 (which is what we get installed right now)
doesn't support Python 3.7 either. :-]