Fixed uploading code coverage information
All attempts to upload code coverage information only resulted in

  Error: No coverage report found

being printed. As it turns out, it appears that automatically locating
all generated report files is disabled in case AppVeyor is detected as
the CI system. However, reading the codecov source code suggests that we
can pass '-f' to explicitely select a file for uploading, so let's do
that.